requirements.txt
fastapi
uvicorn
scikit-learn
joblib
pandas
passlib[bcrypt]
python-jose
pymongo
